KEYCLOAK-11160 - adds "optional" as fallback for conditional OTP
